refactor: Updates markdown config
