+++ lib/extract-posts-from-markdown.ts
@@ -1,9 +1,14 @@
 import fs from "fs";
 import path from "path";
 import matter from "gray-matter";
-import remark from "remark";
-import html from "remark-html";
-import highlight from "remark-highlight.js";
+import unified from "unified";
+import slug from "remark-slug";
+import toc from "remark-toc";
+import remark2rehype from "remark-rehype";
+import highlight from "rehype-highlight";
+import parse from "remark-parse";
+import remark2react from "remark-react";
+import html from "rehype-stringify";
 
 interface Post {
   id: string;
@@ -44,7 +49,20 @@ export function getSortedPostsData() {
 }
 
 export function getAllPostIds() {
+  /**
+   * This function is necessary to extract postIds from the markdown files.
+   * See https://nextjs.org/docs/basic-features/data-fetching#getstaticpaths-static-generation
+   * The returned value should be of the following form:
+   * return {
+   *   paths: [
+   *     { params: { id: '1' } },
+   *     { params: { id: '2' } }
+   *   ],
+   *   fallback: ...
+   * }
+   */
   const fileNames = fs.readdirSync(postsDirectory);
+
   return fileNames.map((fileName) => {
     return {
       params: {
@@ -54,7 +72,7 @@ export function getAllPostIds() {
   });
 }
 
-export async function getPostData(id) {
+export async function getPostData(id: string) {
   const fullPath = path.join(postsDirectory, `${id}.md`);
   const fileContents = fs.readFileSync(fullPath, "utf8");
 
@@ -62,16 +80,20 @@ export async function getPostData(id) {
   const matterResult = matter(fileContents);
 
   // Use remark to convert markdown into HTML string
-  const processedContent = await remark()
+  const processedContent = unified()
+    .use(parse)
+    .use(slug)
+    .use(toc)
+    .use(remark2rehype)
     .use(highlight)
+    .use(remark2react)
     .use(html)
-    .process(matterResult.content);
-  const contentHtml = processedContent.toString();
+    .processSync(matterResult.content).contents;
 
-  // Combine the data with the id and contentHtml
+  // Combine the data with the id and processedContent
   return {
     id,
-    contentHtml,
+    processedContent,
     ...matterResult.data,
   };
 }
